[RPA-1808] Bug with redirecting to registration page. (#807)
* fix: :bug: Fixed issues with redirection to the registration page, invitation code generation, and BackgroundLogo display.

* fix(runbotics-orchestrator-ui): remove unnecessary code

* fix(runbotics-orchestrator-ui): :art:

* refactor(runbotics-orchestrator-ui): :recycle: Refactor after CR

* chore: 📌 bump app version to 2.11.0-SNAPSHOT.87

+++ runbotics/runbotics-scheduler/src/scheduler-database/tenant/tenant.service.ts
@@ -107,8 +107,10 @@ export class TenantService {
             tenantId: tenant.id,
             expirationDate: LessThan(new Date())
         }).then(codes => {
-            this.inviteCodeRepository.delete([...codes.map(code => code.inviteId)]);
-            this.logger.log('Removed expired invite codes for tenant with id: ', tenant.id);
+            if (codes.length > 0) {
+                this.inviteCodeRepository.delete([...codes.map(code => code.inviteId)]);
+                this.logger.log('Removed expired invite codes for tenant with id: ', tenant.id);
+            }
         });
 
         const newInviteCode = {
